refactor event handler
