+++ app/watcher.go
@@ -2,12 +2,9 @@ package app
 
 import (
 	"fmt"
-	"io"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
 
 	"github.com/fsnotify/fsnotify"
 )
@@ -35,27 +32,8 @@ func Watch(params *Params) error {
 				if !ok {
 					return
 				}
-				if eventMatchAsTargetOp(event) {
-					matched, err := doesFileNameMatchFilters(event.Name, *params)
-					if err != nil {
-						log.Fatalf("failed to match file name: %s", err)
-					}
-					if matched {
-						logEvent(event)
-						log.Printf("run command: %v", strings.Join(params.Command, " "))
-						cmd := exec.Command("sh", "-c", strings.Join(params.Command, " "))
-						if err := redirectCommandStdoutAndStderr(cmd); err != nil {
-							log.Fatalf("failed to redirect stdout and stderr: %s", err)
-						}
-
-						if err := cmd.Start(); err != nil {
-							log.Fatalf("failed to run command: %s", err)
-						}
-
-						if err := cmd.Wait(); err != nil && params.ExitOnError {
-							log.Fatalf("failed to wait command: %s", err)
-						}
-					}
+				if err := handleEvent(params, event); err != nil {
+					log.Fatalf("failed to handle event: %s", err)
 				}
 			case err, ok := <-watcher.Errors:
 				if !ok {
@@ -76,48 +54,6 @@ func Watch(params *Params) error {
 	return nil
 }
 
-// eventMatchAsTargetOp checks if event matches as target op.
-func eventMatchAsTargetOp(event fsnotify.Event) bool {
-	return event.Has(fsnotify.Write) || event.Has(fsnotify.Create) || event.Has(fsnotify.Remove)
-}
-
-// redirectCommandStdoutAndStderr redirects stdout and stderr to os.Stdout and os.Stderr.
-func redirectCommandStdoutAndStderr(cmd *exec.Cmd) error {
-	stdout, stderr, err := getStdoutPipeAndStderrPipe(cmd)
-	if err != nil {
-		return fmt.Errorf("failed to get stdout pipe: %w", err)
-	}
-
-	go func() {
-		_, err := io.Copy(os.Stdout, stdout)
-		if err != nil {
-			fmt.Println("failed to copy stdout: %w", err)
-		}
-	}()
-
-	go func() {
-		_, err := io.Copy(os.Stderr, stderr)
-		if err != nil {
-			fmt.Println("failed to copy stderr: %w", err)
-		}
-	}()
-	return nil
-}
-
-// getStdoutPipeAndStderrPipe gets stdout pipe and stderr pipe.
-func getStdoutPipeAndStderrPipe(cmd *exec.Cmd) (io.ReadCloser, io.ReadCloser, error) {
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get stdout pipe: %w", err)
-	}
-
-	stderr, err := cmd.StderrPipe()
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to get stderr pipe: %w", err)
-	}
-
-	return stdout, stderr, nil
-}
 func addDirectoryRecursively(watcher *fsnotify.Watcher, watchDir string) error {
 	if err := filepath.Walk(watchDir, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
