+++ go.mod
@@ -3,9 +3,13 @@ module github.com/akitanak/watcher
 go 1.19
 
 require (
-	github.com/fsnotify/fsnotify v1.6.0 // indirect
+	github.com/fsnotify/fsnotify v1.6.0
+	github.com/spf13/cobra v1.7.0
+	golang.org/x/exp v0.0.0-20230905200255-921286631fa9
+)
+
+require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/cobra v1.7.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sys v0.0.0-20220908164124-27713097b956 // indirect
+	golang.org/x/sys v0.12.0 // indirect
 )
